+++ frontend/src/services/connectionService.jsx
@@ -1,6 +1,10 @@
 import api from './api';
+import { isValidObjectId } from '../utils/validation';
 
 export const sendConnectionRequest = async (receiverId) => {
+  if (!isValidObjectId(receiverId)) {
+    throw new Error('Invalid user ID');
+  }
   const res = await api.post('/connections/request', { receiverId });
   return res.data;
 };
@@ -11,6 +15,9 @@ export const getPendingRequests = async () => {
 };
 
 export const respondToRequest = async (id, action) => {
+  if (!isValidObjectId(id)) {
+    throw new Error('Invalid connection request ID');
+  }
   // action: 'accept' or 'reject'
   const res = await api.put(`/connections/request/${id}`, { action });
   return res.data;
@@ -22,6 +29,9 @@ export const getMyConnections = async () => {
 };
 
 export const removeConnection = async (userId) => {
+  if (!isValidObjectId(userId)) {
+    throw new Error('Invalid user ID');
+  }
   const res = await api.delete(`/connections/${userId}`);
   return res.data;
 };
